+++ src/main/java/de/codescape/jira/plugins/multiplesubtasks/action/UserSubtaskTemplatesAction.java
@@ -20,7 +20,8 @@ public class UserSubtaskTemplatesAction extends JiraWebActionSupport {
 
     private static final long serialVersionUID = 1L;
     private static final long MAXIMUM_TEMPLATES_PER_USER = 10;
-    private static final long MAXIMUM_TEMPLATES_SIZE = 64000;
+    private static final long MAXIMUM_TEMPLATE_LENGTH = 64000;
+    private static final long MAXIMUM_TEMPLATE_NAME_LENGTH = 80;
 
     /**
      * Names of all parameters used on the page.
@@ -96,8 +97,15 @@ public class UserSubtaskTemplatesAction extends JiraWebActionSupport {
     /**
      * Returns the maximum number of characters to be used for a template.
      */
-    public long getMaximumTemplatesSize() {
-        return MAXIMUM_TEMPLATES_SIZE;
+    public long getMaximumTemplateLength() {
+        return MAXIMUM_TEMPLATE_LENGTH;
+    }
+
+    /**
+     * Returns the maximum number of characters to be used for the template name.
+     */
+    public long getMaximumTemplateNameLength() {
+        return MAXIMUM_TEMPLATE_NAME_LENGTH;
     }
 
     /* helper methods */