+++ docs/changelog.md
@@ -30,6 +30,7 @@ Click on the version in the following list to see all changes since the previous
 * feat: ensure label length does not exceed 255 characters
 * feat: support Jira 9.3.1 and enterprise versions 8.20.14 and 8.13.27
 * feat: ensure template names do not exceed 255 characters
+* feat: reduce maximum template name length to 80 characters
 
 ### [22.10.0] - 2022-10-09